Add: 내비게이션에 카트 기능 부착
