+++ src/components/Cart/CartSummary/CartSummary.scss
@@ -30,7 +30,7 @@
     .cartSummaryPayBtn {
       width: 100%;
       padding: 21px 23px;
-      color: black;
+      color: $primary !important;
       background-color: $background;
       font-size: 14px;
     }
